backend/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
requests==2.32.4
boto3==1.36.26
pipecat-ai[aws-nova-sonic,silero]==0.0.74
httpx>=0.28.1
onnxruntime